requirements.txt
crewai>=0.28.0
langchain-huggingface>=0.0.3
faster-whisper>=0.10.0
gTTS>=2.5.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
pygame>=2.5.0
webrtcvad>=2.0.10
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
